[refactor] #4 datastore unit test
