+++ core/datastore/build.gradle.kts
@@ -5,13 +5,9 @@ plugins {
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.serialization)
     alias(libs.plugins.sopt.plugin.test)
-    alias(libs.plugins.sopt.android.test)
 }
 
 android {
-    defaultConfig {
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-    }
     namespace = "org.sopt.datastore"
 }
